Finally fix the weird color glitch
Apparently adding 'Arrays.fill' fixes the glitch. So, the encoder was using the index table of a seen black pixel which is basically garbage that was memset to 0 as the position, but the decoder has actual garbage from previous frame and does not have the black pixel, it just output some random color for that pixel.